updated to nuget
updated all the references in the solution to pull Mvvmcross from nuget.
For the droid library's set the build sdk to 10.  For the Droid
solutions, set the compile with to 14, and the min version to 10.  Also
updated the Xamarin Components, specifically the AppCompat ones to point
to 19.1.  Also added an Android configuration.